code terminé. A peu près rien ne fonctionne

+++ back/controllers/sauces.js
@@ -8,18 +8,23 @@
 */
 
 const sauce = require('../models/sauces');
+const fs = require('fs');
 
 //recevoir et enregistrer une sauce dans la DB
 exports.createSauce = (req, res) => {
-
-    //Supprimer le champ _id que genère mongo db
-    delete req.body._id;
+    const sauceObject = JSON.parse(req.body.sauce);
+    delete sauceObject._id;
+    delete sauceObject._userId;
+                    //new majuscule!!!
     const sauce = new Sauce({
-      ...req.body
-    });
-    sauce.save()
-      .then(() => res.status(201).json({ message :  'objet créé'}))
-      .catch(error => res.status(400).json({error}));
+       ...sauceObject,
+       userId: req.auth.userId,
+       imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+   });
+ 
+   sauce.save()
+   .then(() => { res.status(201).json({message: 'Objet enregistré !'})})
+   .catch(error => { res.status(400).json( { error })});
 };
 
 //récupérer une sauce avec son ID
@@ -31,18 +36,51 @@ exports.getOneSauce = (req, res) => {
 
 //Modifier une sauce 
 exports.modifySauce = (req, res) => {
+    const sauceObject = req.file ? {
+        ...JSON.parse(req.body.sauce),
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+    } : { ...req.body };
+  
+    delete sauceObject._userId;
 
-    //L'ID est redemandé après copie de la requète car elle peut contenir un id différent...
-    Sauce.updateOne({ _id: req.params.id}, {...req.body, _id: req.params.id })
-    .then(() => res.status(200).json({ message: 'Sauce modifiée.'}))
-    .catch(error => res.status(400).json({error}));
+    //majuscule !!!
+    Sauce.findOne({_id: req.params.id})
+        .then((sauce) => {
+            if (sauce.userId != req.auth.userId) {
+                res.status(401).json({ message : 'Not authorized'});
+            } else {
+
+                //majuscule!!!
+                Sauce.updateOne({ _id: req.params.id}, { ...sauceObject, _id: req.params.id})
+                .then(() => res.status(200).json({message : 'Objet modifié!'}))
+                .catch(error => res.status(401).json({ error }));
+            }
+        })
+        .catch((error) => {
+            res.status(400).json({ error });
+        });
 };
 
 //Supprimer une sauce 
 exports.deleteSauce = (req, res) => {
-    Sauce.deleteOne({_id: req.params.id })
-        .then(() => res.status(200).json({ message: 'Sauce supprimée.'}))
-        .catch(error => res.status(404).json({error}));
+    //majuscule
+    Sauce.findOne({ _id: req.params.id})
+       .then(sauce => {
+           if (sauce.userId != req.auth.userId) {
+               res.status(401).json({message: 'Not authorized'});
+           } else {
+               const filename = sauce.imageUrl.split('/images/')[1];
+               fs.unlink(`images/${filename}`, () => {
+                    //majuscule
+                    Sauce.deleteOne({_id: req.params.id})
+                       .then(() => { res.status(200).json({message: 'Objet supprimé !'})})
+                       .catch(error => res.status(401).json({ error }));
+               });
+           }
+       })
+       .catch( error => {
+           res.status(500).json({ error });
+       });
 };
 
 //récupérer toutes les sauces de la DB
